feat: display matching criteria in query patient match list
